ci: builds in release mode (#1076)

+++ Tests/SentryTests/Categories/SentryUnsignedLongLongValueTest.m
@@ -14,13 +14,16 @@
     XCTAssertEqual([@"99" unsignedLongLongValue], 99);
     XCTAssertEqual([@"999" unsignedLongLongValue], 999);
 
-    NSString *longLongMaxValue = [NSString stringWithFormat:@"%ld", 0x7FFFFFFFFFFFFFFF];
+    NSString *longLongMaxValue =
+        [NSString stringWithFormat:@"%llu", (unsigned long long)0x7FFFFFFFFFFFFFFF];
     XCTAssertEqual([longLongMaxValue unsignedLongLongValue], 9223372036854775807);
 
-    NSString *negativelongLongMaxValue = [NSString stringWithFormat:@"%ld", -0x8000000000000000];
+    NSString *negativelongLongMaxValue =
+        [NSString stringWithFormat:@"%llu", (unsigned long long)-0x8000000000000000];
     XCTAssertEqual([negativelongLongMaxValue unsignedLongLongValue], 0x8000000000000000);
 
-    NSString *unsignedLongLongMaxValue = [NSString stringWithFormat:@"%ld", 0xFFFFFFFFFFFFFFFF];
+    NSString *unsignedLongLongMaxValue =
+        [NSString stringWithFormat:@"%llu", (unsigned long long)0xFFFFFFFFFFFFFFFF];
     XCTAssertEqual([unsignedLongLongMaxValue unsignedLongLongValue], 0xFFFFFFFFFFFFFFFF);
 }
